core/serializers.py
from .models import *
from rest_framework import serializers

class TipoProductoSerializer(serializers.ModelSerializer):
    class Meta:
        model = TipoObras
        fields = '__all__'


class ProductoSerializer(serializers.ModelSerializer):
    Tipo = TipoProductoSerializer(read_only=True)

    class Meta:
        model = Producto
        fields = '__all__'

