requirements.txt
django == 4.0.4
pillow
django-crispy-forms
crispy-bootstrap4
psycopg2-binary
django-admin-confirm
django-axes
django-recaptcha
django-simple-captcha
django-rest-framework
whitenoise
cloudinary
reportlab
xhtml2pdf